fix(init): initialization is a publishing concern
Moves the previous init subcommand to a flag in the publish subcommand,
as it is a concern that is particular to publishing and can be
misunderstood easily otherwise.
