+++ src/cli/commands/publish/git/mod.rs
@@ -11,10 +11,10 @@ use super::PublishArgs;
 
 #[derive(Parser, Debug)]
 #[command(next_help_heading = "Git Options")]
-pub(super) struct GitArgs {
+pub(crate) struct GitArgs {
     /// The target remote to publish the atom(s) to
     #[arg(long, short = 't', default_value_t = git::default_remote().to_owned(), name = "TARGET")]
-    remote: String,
+    pub(super) remote: String,
     /// The revision to publish the atom(s) from
     ///
     /// Specifies a revision using Git's extended SHA-1 syntax.
